### 2024-07-25 #7036 @editorialbot
[PRE REVIEW]: osmapiR: An 'OpenStreetMap API' implementation for R
**Submitting author:** @jmaspons (<a href="http://orcid.org/0000-0003-2286-8727">Joan Maspons</a>)
**Repository:** https://github.com/ropensci/osmapiR
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.0
