### 2024-07-25 #7036 @arfon
@jmaspons - many thanks for your submission to JOSS. It looks like the paper isn't currently compiling, could you please take a look at this example paper https://joss.readthedocs.io/en/latest/example_paper.html for the exact formatting requirements and update your paper accordingly (both metadata/YAML but also the structure/content)?
