### 2024-07-25 #7036 @jmaspons
Works with the gh action in my repo https://github.com/ropensci/osmapiR/actions/runs/10090439427

> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/10097625931?check_suite_focus=true).

Seems a problem with the (lack of) affiliation. I will try to fix it
