### 2024-07-26 fac3a77
Add affiliation as per https://github.com/openjournals/joss/issues/406
