### 2024-07-26 #7036 @jmaspons
Seems related to https://github.com/openjournals/joss/issues/406. Hopefully fixed with the workaround
