### 2024-08-09 33e4640
Fix case with only one key
