### 2024-08-10 9d92113
Defensive changes
Check for type id columns in osmchange_*()
Avoid simplification of 1 col data frames
