### 2024-08-28 #7036 @arfon
Related rOpenSci review here: https://github.com/ropensci/software-review/issues/633
