### 2024-08-28 #7036 @arfon
:wave: @jonthegeek, @ccamara – there's no additional review work to do here, rather, we're crediting you for your efforts on the earlier rOpenSci review: https://github.com/ropensci/software-review/issues/633
