### 2024-09-02 bb22813
Prepare release 0.2.0 for JOSS in 10.5281/zenodo.13627998
