+++ README.Rmd
@@ -23,6 +23,7 @@ knitr::opts_chunk$set(
 [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/osmapiR)](https://cran.r-project.org/package=osmapiR)
 [![CRAN Downloads](https://cranlogs.r-pkg.org/badges/grand-total/osmapiR)](https://cran.r-project.org/package=osmapiR)
 [![](https://badges.ropensci.org/633_status.svg)](https://github.com/ropensci/software-review/issues/633)
+[![status](https://joss.theoj.org/papers/eed311474b650ed64042431a537fbc35/status.svg)](https://joss.theoj.org/papers/eed311474b650ed64042431a537fbc35)
 <!-- badges: end -->
 
 An R interface to the [OpenStreetMap API v0.6](https://wiki.openstreetmap.org/wiki/API_v0.6) for fetching and saving raw
