### 2024-09-03 0368f1b
Test and fix tags_list2wide() with only 1 tag per object
Also add tests for tags_wide2list()
