### 2024-09-04 #7151 @arfon
@jonthegeek, @ccamara – many thanks for your (rOpenSci) reviews. JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
